moved Request and P2P headers into lower level of library, confirmed build,
