+++ include/CMakeLists.txt
@@ -2,3 +2,5 @@ target_include_directories(MPIPCL PUBLIC
 	$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
 	$<INSTALL_INTERFACE:include>
 )
+
+message(STATUS "TEST")
