+++ src/core/CMakeLists.txt
@@ -1,1 +1,7 @@
+
 add_subdirectory(pt2pt)
+
+target_include_directories(MPIPCL PUBLIC
+	$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+	$<INSTALL_INTERFACE:/src./core/include>
+)
