+++ src/core/include/CMakeLists.txt
@@ -1,0 +1,7 @@
+message(STATUS "TEST1")
+
+#target_include_directories(MPIPCL PUBLIC
+#	$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+#)
+
+message(STATUS "TEST2")
